startup services

+++ users/api/controllers/deleteUser.js
@@ -3,29 +3,26 @@ const db = require('../utils/postgres');
 /* DELETE User  */
 module.exports = async (req, res) => {
   try {
-    const { rows } = await db.query(
+    const user = await db.query(
       `SELECT * FROM users WHERE id = '${req.params.id}'`,
     );
-    rows.length !== 1 &&
+
+    if (user.rows.length === 1) {
+      //TODO: rajouter une condition sur l'id du token = id en paramètre OR Role = true (admin)
+      const { rows } = await db.query(
+        `DELETE FROM users WHERE id = '${req.params.id}'`,
+      );
+
+      res.status(204).json({
+        status: 'Success',
+        message: 'User Deleted',
+      });
+    } else {
       res.status(404).json({
         status: 'Error',
         message: 'User not found',
       });
-  } catch (err) {
-    res.status(500).json({
-      status: 'Server Internal Error',
-      message: error,
-    });
-  }
-  try {
-    //TODO: rajouter une condition sur l'id du token = id en paramètre OR Role = true (admin)
-    const { rows } = await db.query(
-      `DELETE FROM users WHERE id = '${req.params.id}'`,
-    );
-
-    res.status(204).json({
-      status: 'Success',
-    });
+    }
   } catch (error) {
     res.status(500).json({
       status: 'Server Internal Error',
